fix: unconditionally parse signing token from JWT_SECRET
When OAuth is not enabled, the general token signing secret is not
parsed, resulting in a different secret per process. This can render
values signed by one process invalid when read by another process.

Fix this by always parsing the OAuth JWT_SECRET, regardless of whether
OAuth is enabled or not, in order to populate the general token
signing secret for other components.